Implementation of PurchaseManager and LinearDutchAuction contracts:
PurchaseManager provides the `managePurchase(uint)` modifier to handle all
per-address, per-tx, and total-inventory limits. It also handles costing,
revenues, and reimbursement for overpaying if the purchased amount is capped.

The virtual function cost(uint n) is expected to return the cost of purchasing
n items at that point in time. The LinearDutchAuction contract is implemented
by overriding this function with a linearly decreasing return value.

TODO: thorough tests are in place for all but flow of funds, which need to be
added before use in production. Committing early to back up a significant
amount of new code.