DynamicBuffer.allocate() no longer returns the container element.
Per discussion with cxkoda, the container element is not actually used and
can always be recovered by decrementing the memory position of the buffer.
It does however, add a potential bug surface for users of the library as
they may attempt to append bytes to the container instead of the buffer.