Introduce ethtest.SimulatedBackend.Must()
Many transactions in tests are performed merely to achieve a required internal state,
and not for actually testing functionality. The boilerplate of `if _, err := Foo(); err != nil`
adds unnecessary bloat and distracts from test logic, adding risk that a bug is missed.

Note that this is *not* to be considered precedent for introducing assert/require libraries
into ethier and regular Go testing MUST be preferred unless a specific use case can be
demonstrated. We can decide on such introductions on a case-by-case basis; prefer differs
like errdiff to functions that directly report on testing.T.