Initial implementation of a fair off-chain address raffle.
The list of entrants itself as well as an independent entropy source are both
used to seed a shuffle of the full entrant list, from which the first n are
returned. Seeding uses keccak256 hashes of each of the two sources; this results
in 2x256 = 2x4x64-bit words that are xor'd and used as a seed to math/rand.

The use of the entrant list as a source of entropy allows for the reuse of the
independent source should it be expensive to generate (e.g. Chainlink VRF).
Different raffles are guaranteed to have different results as long as their
list of entrants is not identical.