OpenSeaERC721Mintable contract for minting directly on OpenSea.
Context: https://docs.opensea.io/docs/2-custom-item-sale-contract

This is a more abstract implementation of the example shown in the OpenSea
Creatures repository, allowing for reuse via inheritance. Furthermore, the
existence of a separate factory contract and its associated behaviour are
all abstracted away so the user is oblivious to its existence.