Chainlink chain-agnostic deployment with built-in constants + Go test package.
The VRFConsumerHelper contract acts as a drop-in replacement for the standard
VRFConsumerBase from Chainlink, but doesn't require specification of constants
like fees, contract addresses, and key hashes. This is benchmarked and adds
<1000 gas units per request, compared to the standard approach, but without the
risk of deployment with an incorrect address.

The helper also has a withdrawLINK() function to withdraw tokens belonging to
the contract.

A set of test-double fakes of the LINK token and a VRFCoordinator are also
included, along with the Go chainlinktest package that deploys them to a
geth SimulatedBackend. The VRFConsumerHelper is also compatible with these
contracts to allow for complete VRF testing with ethier.