PaymentSplitterDeployer for EIP-1167 PaymentSplitter deployments (#24)
* PaymentSplitterDeployer for cheap PaymentSplitter deployments with minimal contract proxies.

Deployment of an OpenZeppelin PaymentSplitter costs at least 1.3M gas. This commit introduces
a slightly modified version, ProxyPaymentSplitter, derived from the original PaymentSplitter,
with the constructor changed to an init function——this makes it amenable to DELEGATECALL usage
and the EIP-1167 minimal contract proxy.

The PaymentSplitterDeployer is intended for single deployment on each EVM-compatible chain. It
deploys a primary instance of the ProxyPaymentSplitter and exposes deploy() and
deployDeterministic() functions for deploying proxies to the primary instance. The deployer
immediately calls the init() function so as to mimic standard PaymentSplitter deployment.

As the ProxyPaymentSplitter shares an interface with PaymentSplitter (other than the constructor
and init functions) it can be used wherever a PaymentSplitter is expected, even in already-
deployed contracts.

An example deployment shows savings of ~0.130Ξ per deployment at a gas price of 100 gwei, which
is approximately $450 saved per deployment at an ETH price of $3500. Both of these values are
reasonable in today's ETH mainnet environment.

The SimulatedBackend is also updated to produce deterministic addresses so sas to not break the
gas-saving testable example.

* EIP-1167 minimal proxy PaymentSplitter: use @openzeppelin/contracts-upgradeable base.

The Solidity code is also moved out of ./contracts and into ./factories to signal that
it isn't expected to be used directly.