Implement ERC721Redeemer library for redeeming claiming rights afforded by an ERC721.
Holders of NFTs may be granted rights to arbitrary actions, with a maximum number per token.
The ERC721Redeemer.Claims struct has redeem() and unclaimed() functions that implement this
logic, limiting to a maxAllowance per token regardless of transfers. Both the token owner &
the approved address of the token can make a claim, but only once.