Introduce ArbitraryPriceSeller with breaking change to Seller.cost() signature.
Seller requires a user to implement a cost(n) function that returns the cost of n
purchases. A second metadata parameter is added so the signature is cost(uint,uint)
and the metadata is propagated via Seller._purchase(), which also accepts this new
parameter. A convenience version of _purchase() with the old signature merely passes
0 as the metadata.

This pattern allows arbitrary data (not necessarily price) to be propagated from the
caller of _purchase() to the cost() function, for use in determining the cost. This
could be, for example, a fixed discount.

The ArbitraryPriceSeller implements a cost() function for which the metadata is the
cost per item. A call to _purchase(to,n,0.5 ether) will therefore cost n/2 ether in
total. The _purchase() convenience function will always revert with an
ArbitraryPriceSeller to avoid accidental free purchases.