Implement SignerFromPRF() method on eth.HDPathPrefix.
A pseudorandom function (PRF) is a cryptographically secure means of deterministically
generating entropy that is indistinguishable from random for anyone without access to
the function's key. It can therefore be used to compute a set of different HD wallets
with only one key to be protected. Although each will have an associated mnemonic, the
intended use case for SignerFromPRF() is in automated systems.

Google's Tink packages are used for PRF primitives and key management.