wETH chain-agnostic library for token address + Go test library.
The WETH library determines the address of the wETH contract based on the current
chain ID and also provides an interface for interaction. This is coupled with the
wethtest Go package that deploys a simulated wETH contract when using ethier's
ethtest.SimulatedBackend. The library and the package are compatible with one
another such that the library returns the address of the contract deployed by the
wethtest package.