*BREAKING CHANGE* Move call of _handlePurchase() in Seller._purchase() to be an "interaction".
As a purchase is often an ERC721 safeMint() it constitutes an interaction. Although the function
has the nonReentrant modifier, still adhering to checks-effects-interactions is a safety measure
in case the modifier is later removed.

Note: this introduces a breaking change in that the value of Seller._totalSold() will now be the
post-purchase count whereas before it was documented as being the pre-purchase count.