Experimental ethier binary UI for managing seed phrases of hot wallets.
The initial intended use case is to create secure backups of seed phrases for already-hot
wallets. Hardware-wallet recovery phrases MUST NOT be used with this feature.

Assumes that GCP default credentials already exist when running, which can be created
with `gcloud auth application-default login` from the GCP SDK. Adding support for AWS KMS
and HashiCorp Vault is trivial and can be done later.

This commit doesn't include tests yet because I'm not quite sure how to test with the
prompts. I think the best option is to have injectable functions that perform prompts and
return user "desires", which can be stubbed out in testing.