Trace-based coverage reports.
Unlike instrumentation-based coverage, trace-based analysis leaves code intact. This guarantees that production and test code are identical; therefore gas consumption and contract size are both reflective are reality.

The output is an LCOV trace file (see `man geninfo`) that can be rendered with [`genhtml`](https://github.com/linux-test-project/lcov/blob/master/bin/genhtml) but also has more widespread support as a standard format. Currently only line hits are recorded; branch and function hits are planned on the roadmap.

The current implementation adds a bit of bloat to the generated code, especially with imported libraries, but this can be reduced in future iterations.