chore: pin the Bazel version
This ensures that all developers have the same version of bazel as the CI machine.
Especially important if we decide to upgrade the version used, so devs aren't broken by missing features.
Note, Bazelisk, as well as Bazel's wrapper script you get from the installer, will check this file to prevent a wrong Bazel from running.
If we want a stronger guarantee, we can follow
https://github.com/bazelbuild/bazelisk#ensuring-that-your-developers-use-bazelisk-rather-than-bazel